cloud/pyproject.toml
[tool.poetry]
name = "cloud"
version = "0.1.0"
description = ""
authors = ["B4rtware <34386047+B4rtware@users.noreply.github.com>", "Re-Krass <38668040+Re-Krass@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyzmq = "^23.1.0"
fastapi = "^0.78.0"
uvicorn = "^0.18.1"
pydantic = "^1.9.1"
SQLAlchemy = "^1.4.39"
tornado = "^6.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
sqlalchemy2-stubs = "^0.0.2-alpha.24"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
